Drive arcade and Joystick deadzones
Nick worked with Mentor Michael to add a drive arcade system to the drive train. This system is now active in the DriveWithJoy command instead of drive by tank.

Issac worked with Mentor Michael to debug drive arcade, and to add a deadzone filter to the IO joysticks to prevent inputs being sent through the x-box joysticks when they are not being manipulated.